+++ CHANGELOG.rst
@@ -34,6 +34,8 @@ Internal Changes
 - Added Python 3.7 and Python 3.8 to the CI. Use the latest version of Python 3
   for development. (:issue:`21`, :pr:`189`). `Aaron Spring`_
 - Lint with the latest black. (:issue:`179`, :pr:`191`). `Ray Bell`_
+- Update mape algorithm from scikit-learn v0.24.0 and test against it.
+  (:issue:`160`, :pr:`230`) `Ray Bell`_
 
 
 xskillscore v0.0.18 (2020-09-23)